backend/requirements.txt
# Web Framework et ASGI
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# Base de données et ORM
sqlalchemy>=2.0.0
alembic>=1.11.0
# Drivers de base de données
psycopg2-binary>=2.9.6  # PostgreSQL
pymysql>=1.0.2  # MariaDB/MySQL
aiosqlite>=0.19.0  # SQLite async

#GRAPHQL
strawberry-graphql[fastapi]>=0.278.0
strawberry-sqlalchemy
strawberry-sqlalchemy-mapper
strawchemy

#Gestion des tâches
celery>=5.5.3
redis>=4.2.0
setuptools>=67.7.2  # Nécessaire pour installer des paquets avec des dépendances C

# HTTP et WebSocket
httpx>=0.24.0
websockets>=11.0.3

# Stockage et Index de recherche
whoosh-reloaded >=2.7.4
tinydb>=4.8.0
cachetools>=5.3.2

# Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
aiofiles>=23.1.0
pathlib>=1.0.1

# Sécurité
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Tests
pytest>=7.3.1
pytest-asyncio>=0.21.0